move title_format to container
